+++ src/Blocky.php
@@ -3,10 +3,14 @@
 namespace wrux\blocky;
 
 use Craft;
+use craft\base\BlockElementInterface;
 use craft\base\Plugin;
+use craft\elements\db\MatrixBlockQuery;
 use craft\web\twig\variables\CraftVariable;
 use wrux\blocky\twig\BlocksTwigExtension;
 use yii\base\Event;
+use wrux\blocky\BlockParser;
+use wrux\blocky\exceptions\BlockTransformerNotFoundException;
 use wrux\blocky\variables\BlockParserVariable;
 
 /**
@@ -24,9 +28,9 @@ class Blocky extends Plugin {
   /**
    * Plugin singelton instance
    *
-   * @var BlockParser
+   * @var Blocky
    */
-  public static $plugin;
+  public static Blocky $plugin;
 
   // Public Methods
   // ===========================================================================
@@ -55,5 +59,54 @@ class Blocky extends Plugin {
 
     // Add the blocks twig tag.
     Craft::$app->view->registerTwigExtension(new BlocksTwigExtension());
+  }
+
+  /**
+   * Parse the Matrix blocks array
+   *
+   * @param array|MatrixBlockQuery $blocks
+   * @return void
+   */
+  public function parseBlocks($blocks) {
+    // If the matrix block was not egar loaded then execute the query.
+    if ($blocks instanceof MatrixBlockQuery) {
+      $blocks = $blocks->all();
+    }
+    if (!$blocks || count($blocks) === 0) {
+      return [];
+    }
+    $block_parser = new BlockParser;
+    foreach ($blocks as $block) {
+      if (!$block instanceof BlockElementInterface) {
+        continue;
+      }
+      try {
+        $block_parser->addBlock($block);
+      }
+      catch  (BlockTransformerNotFoundException $e) {
+        // The array contains data for a block with no corresponding parser class.
+        Craft::warning(
+          Craft::t(
+            'blocky',
+            'Block transformer not found: {erorr}',
+            ['error' => $e->getMessage()]
+          ),
+          __METHOD__
+        );
+        continue;
+      }
+      catch (\Exception $e) {
+        Craft::error(
+          Craft::t(
+            'blocky',
+            'There was an error parsing the blocks: {erorr}',
+            ['error' => $e->getMessage()]
+          ),
+          __METHOD__
+        );
+        return [];
+      }
+    }
+    return $block_parser;
   }
 }